Test 1 Modifica y Elimina Usuarios sin pruebas

+++ src/main/java/gei/id/tutelado/dao/UsuarioDaoJPA.java
@@ -65,10 +65,10 @@ public class UsuarioDaoJPA implements UsuarioDao {
 			em = emf.createEntityManager();
 			em.getTransaction().begin();
 
-			int numPruebas;
+			Long numPruebas;
 			Usuario usuarioTmp = em.find(Usuario.class, usuario.getId());
 			if (usuarioTmp.esSanitario()) {
-				numPruebas = em.createNamedQuery("Sanitario.recuperarNumPruebas", Integer.class).setParameter("dni", usuario.getDni()).getSingleResult();
+				numPruebas = em.createNamedQuery("Sanitario.recuperarNumPruebas", Long.class).setParameter("dni", usuario.getDni()).getSingleResult();
 				if (numPruebas > 0) {
 					throw new RuntimeException("¡El sanitario ha realizado pruebas, no se puede eliminar!");
 				}
